Move dialogue.py into RimSort.models
