+++ RimSort/windows/missing_mods_panel.py
@@ -1,32 +1,21 @@
 from functools import partial
 from loguru import logger
-import os
-import platform
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict
 
-
-from PySide6.QtCore import Qt, QEvent, QModelIndex, QObject, QPoint, QSize, Signal
-from PySide6.QtGui import QAction, QPixmap, QStandardItemModel, QStandardItem
+from PySide6.QtCore import Qt, QEvent, QSize, Signal
+from PySide6.QtGui import QStandardItemModel, QStandardItem
 from PySide6.QtWidgets import (
     QAbstractItemView,
     QComboBox,
-    QGridLayout,
     QHeaderView,
-    QItemDelegate,
     QLabel,
-    QListWidget,
-    QListWidgetItem,
-    QMenu,
     QPushButton,
-    QSizePolicy,
     QTableView,
-    QToolButton,
     QWidget,
     QHBoxLayout,
     QVBoxLayout,
 )
 
-from model.dialogue import show_warning
 from util.constants import RIMWORLD_DLC_METADATA
