+++ model/mod_list.py
@@ -19,7 +19,11 @@ from PySide6.QtWidgets import (
 
 from RimSort.controllers.settings_controller import SettingsController
 from model.mod_list_item import ModListItemInner
-from model.dialogue import show_dialogue_conditional, show_dialogue_input, show_warning
+from RimSort.models.dialogue import (
+    show_dialogue_conditional,
+    show_dialogue_input,
+    show_warning,
+)
 from RimSort.utils.event_bus import EventBus
 from util.generic import (
     delete_files_except_extension,
