+++ util/schema.py
@@ -1,7 +1,7 @@
 from loguru import logger
 from typing import Any, Dict
 
-from model.dialogue import show_warning
+from RimSort.models.dialogue import show_warning
 
 
 def validate_mods_config_format(mods_config_data: Dict[str, Any]) -> bool:
