+++ util/steam/browser.py
@@ -24,7 +24,7 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
 )
 
-from model.dialogue import show_warning
+from RimSort.models.dialogue import show_warning
 from RimSort.models.image_label import ImageLabel
 from util.steam.webapi.wrapper import (
     ISteamRemoteStorage_GetCollectionDetails,
